Moved the external_warning_formatting method to exceptions.py and updated it.
